Create a wrapper component for ReactJSON
